Huge refactor of multiple systems

+++ core/src/main/java/com/karatesan/game/ecs/components/economy/XpComponent.java
@@ -4,10 +4,10 @@ import com.badlogic.ashley.core.Component;
 import com.badlogic.gdx.utils.Pool.Poolable;
 
 public class XpComponent implements Component, Poolable {
-    public float value = 1f;
+    public int value = 1;
 
     @Override
     public void reset() {
-        value = 1f;
+        value = 1;
     }
 }
